+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unpublished
 
 - Add better `panic!` message to `map_range` if cast fails
+- Add many items to prelude (audio, io, math, osc, ui, window).
 
 # Version 0.6.0 (2017-06-07)